refactor: simplify project architecture
