+++ packages/runner/src/execution/run-scenario.ts
@@ -1,150 +1,43 @@
-import {
-  ignoreTag,
-  type ScenarioStep,
-  type ScenarioVariableBinding,
-} from '@pickle-spec/spec'
-import { persistedEvidenceKinds } from '../evidence/evidence'
-import {
-  cachedStepPrefixFrom,
-  evaluationAt,
-  type GapCursor,
-  gapCursor,
-  reseatGap,
-} from '../execution-cache/cached-step-prefix'
-import type { ExecutionCacheEnvelope } from '../execution-cache/execution-cache'
-import { requiredValue } from '../required-value'
+import { ignoreTag } from '@pickle-spec/spec'
 import { withSharedEvidenceObservations } from '../results/shared-evidence-observations'
 import type {
-  ActionEvidence,
-  ActionEvidenceInput,
-  DiagnosticEntry,
-  EvidenceAvailability,
-  ExecutionCachePolicy,
-  ExecutionMode,
-  ExecutionTimeouts,
+  AttemptScenarioRun,
   RunEvent,
-  RunEventPayload,
-  RunEventScope,
-  RunScenarioInput,
   ScenarioAttempt,
-  ScenarioRun,
-  ScenarioTargetSession,
-  StepEvaluation,
+  ScenarioAttemptInput,
   StepExecution,
-  StepTargetSession,
   TargetSession,
   TargetSessionCompletion,
-  TestResult,
   TestResultState,
   TestStepResult,
 } from './run-scenario-types'
-import { evidenceKinds, testRunSchemaVersion } from './run-scenario-types'
+import { testRunSchemaVersion } from './run-scenario-types'
+import { recordExecutionError } from './scenario/scenario-diagnostics'
 import {
-  nonemptyBindings,
-  publicActionEvidence,
-  publicStepExecution,
-  redactString,
-  scenarioDefinitionId,
-  scenarioIdentity,
-  stringContainsBinding,
-  templateStepAt,
-} from './scenario-runtime'
+  type AttemptProgress,
+  attemptIdentity,
+  type EmitAttemptEvent,
+  type RecordStep,
+} from './scenario/scenario-execution-context'
+import {
+  attemptEvidence,
+  createTestResult,
+  durationMs,
+  scenarioFinishedPayload,
+} from './scenario/scenario-results'
+import { nonemptyBindings } from './scenario/scenario-runtime'
+import {
+  executeTargetSession,
+  recordStepExecution,
+} from './scenario/scenario-target-session'
 
 export * from './run-scenario-types'
-
-class ExecutionDeadlineError extends Error {
-  constructor(message: string) {
-    super(message)
-    this.name = 'ExecutionDeadlineError'
-  }
-}
-
-function errorMessage(error: unknown): string {
-  return error instanceof Error ? error.message : String(error)
-}
-
-function isCancellation(error: unknown, signal?: AbortSignal): boolean {
-  return (
-    Boolean(signal?.aborted) ||
-    (error instanceof Error && error.name === 'AbortError')
-  )
-}
-
-function validateCompletion(
-  completion: TargetSessionCompletion,
-): TargetSessionCompletion {
-  if (
-    !Number.isSafeInteger(completion.inferenceCount) ||
-    completion.inferenceCount < 0
-  ) {
-    throw new Error(
-      'Target session completion requires a non-negative integer inferenceCount',
-    )
-  }
-  return completion
-}
-
-function executeWithDeadline<T>(
-  operation: (signal?: AbortSignal) => Promise<T>,
-  signal: AbortSignal | undefined,
-  timeoutMs: number | undefined,
-  timeoutMessage: string,
-): Promise<T> {
-  if (timeoutMs === undefined) return operation(signal)
-
-  const controller = new AbortController()
-  const onAbort = () => controller.abort()
-  signal?.addEventListener('abort', onAbort, { once: true })
-
-  return new Promise<T>((resolve, reject) => {
-    const timer = setTimeout(() => {
-      controller.abort()
-      reject(new ExecutionDeadlineError(timeoutMessage))
-    }, timeoutMs)
-    operation(controller.signal)
-      .then(resolve, reject)
-      .finally(() => {
-        clearTimeout(timer)
-        signal?.removeEventListener('abort', onAbort)
-      })
-  })
-}
-
-function stepDeadline(
-  timeout: ExecutionTimeouts | undefined,
-  scenarioStartedAt: number,
-): { timeoutMs: number | undefined; timeoutMessage: string } {
-  const scenarioRemaining =
-    timeout?.scenarioMs === undefined
-      ? undefined
-      : timeout.scenarioMs - (Date.now() - scenarioStartedAt)
-  const usesScenarioDeadline =
-    scenarioRemaining !== undefined &&
-    (timeout?.stepMs === undefined || scenarioRemaining <= timeout.stepMs)
-  return {
-    timeoutMs: usesScenarioDeadline
-      ? Math.max(0, scenarioRemaining)
-      : timeout?.stepMs,
-    timeoutMessage: usesScenarioDeadline
-      ? `Scenario exceeded its ${timeout?.scenarioMs}ms deadline`
-      : `Step exceeded its ${timeout?.stepMs}ms deadline`,
-  }
-}
-
-export interface ScenarioAttemptInput extends RunScenarioInput {
-  mode: ExecutionMode
-  attempt: number
-  cacheEntry?: ExecutionCacheEnvelope
-}
-
-export interface AttemptScenarioRun extends ScenarioRun {
-  attempt: ScenarioAttempt
-  completion?: TargetSessionCompletion
-  replayDiverged: boolean
-  runtimeValueExposed: boolean
-  replayedStepCount: number
-  adaptiveEvaluated: boolean
-}
+export {
+  createSyntheticTestResult,
+  createTestResult,
+  scenarioFinishedPayload,
+  withFinalAttempt,
+} from './scenario/scenario-results'
 
 interface InitialAttemptOutcome {
   state: 'skipped' | 'cancelled'
@@ -164,239 +57,6 @@ function initialAttemptOutcome(
     }
   }
   return undefined
-}
-
-function stampDiagnostic(
-  entry: DiagnosticEntry,
-  input: ScenarioAttemptInput,
-  stepIndex?: number,
-  step?: ScenarioStep,
-): DiagnosticEntry {
-  return {
-    ...entry,
-    scenarioId:
-      entry.scenarioId ??
-      scenarioDefinitionId(input.specification, input.scenario),
-    scenarioName: entry.scenarioName ?? input.scenario.name,
-    stepIndex: entry.stepIndex ?? stepIndex,
-    stepText:
-      entry.stepText ??
-      (step ? `${step.keyword.trim()} ${step.text}` : undefined),
-    executionTargetProfileId:
-      entry.executionTargetProfileId ?? input.executionTargetProfile.id,
-  }
-}
-
-function runnerDiagnostic(
-  input: ScenarioAttemptInput,
-  message: string,
-  occurredAt: string,
-  stepIndex?: number,
-  step?: ScenarioStep,
-): DiagnosticEntry {
-  return stampDiagnostic(
-    {
-      occurredAt,
-      level: 'error',
-      origin: 'runner',
-      message,
-    },
-    input,
-    stepIndex,
-    step,
-  )
-}
-
-const evidenceCapabilities = {
-  screenshot: 'screenshots',
-  trace: 'traces',
-  recording: 'recordings',
-  'device-log': 'device-logs',
-  diagnostics: 'diagnostics',
-} as const
-
-function attemptEvidence(
-  input: ScenarioAttemptInput,
-  steps: readonly TestStepResult[],
-  reported: readonly EvidenceAvailability[] = [],
-  diagnostics: readonly DiagnosticEntry[] = [],
-): EvidenceAvailability[] {
-  const available = persistedEvidenceKinds(steps, diagnostics)
-  const capabilities = new Set(
-    input.executionTargetProfile.capabilities ?? input.adapter.capabilities,
-  )
-  return evidenceKinds.map((kind) => {
-    if (available.has(kind)) {
-      return { kind, state: 'available' }
-    }
-    const adapterAvailability = reported.findLast(
-      (availability) => availability.kind === kind,
-    )
-    return (
-      adapterAvailability ?? {
-        kind,
-        state: capabilities.has(evidenceCapabilities[kind])
-          ? 'not-requested'
-          : 'not-supported',
-      }
-    )
-  })
-}
-
-export function createTestResult(
-  input: ScenarioAttemptInput,
-  attempts: ScenarioAttempt[],
-): TestResult {
-  const first = attempts[0]
-  const final = attempts.at(-1)
-  if (!first || !final) {
-    throw new Error('A Test result requires at least one Scenario attempt')
-  }
-  const scenarioId = scenarioDefinitionId(input.specification, input.scenario)
-  return {
-    schemaVersion: testRunSchemaVersion,
-    specification: {
-      name: input.specification.name,
-      uri: input.specification.source.uri,
-    },
-    scenario: { ...scenarioIdentity(input.scenario), id: scenarioId },
-    executionTargetProfile: input.executionTargetProfile,
-    state: final.state,
-    startedAt: first.startedAt,
-    finishedAt: final.finishedAt,
-    durationMs: durationMs(first.startedAt, final.finishedAt),
-    attempts,
-    ...(attempts.length > 1 && final.state === 'passed' ? { flaky: true } : {}),
-  }
-}
-
-export function createSyntheticTestResult(
-  input: RunScenarioInput,
-  mode: ExecutionMode,
-  state: TestResultState,
-  message?: string,
-): TestResult {
-  const occurredAt = (input.now ?? (() => new Date()))().toISOString()
-  const attemptInput: ScenarioAttemptInput = {
-    ...input,
-    mode,
-    attempt: 1,
-  }
-  const attempt: ScenarioAttempt = {
-    attempt: 1,
-    startedAt: occurredAt,
-    finishedAt: occurredAt,
-    durationMs: 0,
-    state,
-    steps: [],
-    executionMode: mode,
-    inferenceCount: 0,
-    ...(input.adapter.fidelityPolicy
-      ? { fidelityPolicy: input.adapter.fidelityPolicy }
-      : {}),
-    ...(message !== undefined ? { message } : {}),
-    evidenceAvailability: attemptEvidence(attemptInput, []),
-  }
-  return createTestResult(attemptInput, [attempt])
-}
-
-export function withFinalAttempt(
-  result: TestResult,
-  update: Partial<ScenarioAttempt>,
-): TestResult {
-  const attempts = result.attempts.map((attempt, index) =>
-    index === result.attempts.length - 1 ? { ...attempt, ...update } : attempt,
-  )
-  const final = requiredValue(attempts.at(-1))
-  return {
-    ...result,
-    state: final.state,
-    finishedAt: final.finishedAt,
-    durationMs: durationMs(result.startedAt, final.finishedAt),
-    attempts,
-    flaky: attempts.length > 1 && final.state === 'passed' ? true : undefined,
-  }
-}
-
-export function scenarioFinishedPayload(result: TestResult): RunEventPayload {
-  const attempt = result.attempts.at(-1)
-  if (!attempt) throw new Error('A Test result requires a Scenario attempt')
-  return {
-    type: 'scenario-finished',
-    specification: result.specification,
-    scenario: result.scenario,
-    executionTargetProfile: result.executionTargetProfile,
-    scope: {
-      scenarioId: requiredValue(result.scenario.id),
-      examplesRowId: result.scenario.examplesRowId,
-      executionTargetProfileId: result.executionTargetProfile.id,
-      attempt: attempt.attempt,
-    },
-    attempt,
-  }
-}
-
-function attemptIdentity(input: ScenarioAttemptInput, stepIndex?: number) {
-  const scenario = {
-    ...scenarioIdentity(input.scenario),
-    id: scenarioDefinitionId(input.specification, input.scenario),
-  }
-  const scope: RunEventScope = {
-    scenarioId: scenario.id,
-    examplesRowId: scenario.examplesRowId,
-    executionTargetProfileId: input.executionTargetProfile.id,
-    attempt: input.attempt,
-    stepIndex,
-  }
-  return {
-    scenario,
-    executionTargetProfile: input.executionTargetProfile,
-    scope,
-  }
-}
-
-function durationMs(startedAt: string, finishedAt: string): number {
-  return Math.max(0, Date.parse(finishedAt) - Date.parse(startedAt))
-}
-
-interface AttemptProgress {
-  state: TestResultState
-  message?: string
-  replayDiverged: boolean
-  runtimeValueExposed: boolean
-  evidenceAvailability: EvidenceAvailability[]
-  diagnostics: DiagnosticEntry[]
-  replayedStepCount: number
-  adaptiveEvaluated: boolean
-}
-
-type EmitAttemptEvent = (
-  event: RunEventPayload,
-  occurredAt?: string,
-) => Promise<RunEvent>
-type RecordExecution = (
-  stepIndex: number,
-  startedAt: string,
-  execution: StepExecution,
-) => Promise<boolean>
-type UntimedTestStepResult = Omit<
-  TestStepResult,
-  'index' | 'startedAt' | 'finishedAt' | 'durationMs'
->
-type RecordStep = (
-  stepIndex: number,
-  startedAt: string,
-  result: UntimedTestStepResult,
-) => Promise<void>
-
-interface SessionExecutionContext {
-  input: ScenarioAttemptInput
-  bindings: readonly ScenarioVariableBinding[]
-  progress: AttemptProgress
-  latestOccurredAt: () => string
-  emit: EmitAttemptEvent
-  recordExecution: RecordExecution
-  recordStep: RecordStep
 }
 
 interface FinishAttemptInput {
@@ -453,419 +113,6 @@ async function finishAttempt(
     replayedStepCount: context.progress.replayedStepCount,
     adaptiveEvaluated: context.progress.adaptiveEvaluated,
   }
-}
-
-interface RecordExecutionInput {
-  input: ScenarioAttemptInput
-  bindings: readonly ScenarioVariableBinding[]
-  progress: AttemptProgress
-  recordStep: RecordStep
-}
-
-async function recordStepExecution(
-  context: RecordExecutionInput,
-  stepIndex: number,
-  startedAt: string,
-  execution: StepExecution,
-): Promise<boolean> {
-  const progress = context.progress
-  const templateStep = templateStepAt(context.input.scenario, stepIndex)
-  const projected = publicStepExecution(execution, context.bindings)
-  progress.runtimeValueExposed ||= projected.runtimeValueExposed
-  progress.evidenceAvailability.push(
-    ...(projected.execution.evidenceAvailability ?? []),
-  )
-  if (context.input.signal?.aborted) {
-    progress.state = 'cancelled'
-    progress.message = 'Scenario cancelled during step execution'
-    await context.recordStep(stepIndex, startedAt, {
-      step: templateStep,
-      state: progress.state,
-      resolvedActions: projected.execution.resolvedActions,
-      message: progress.message,
-    })
-    return false
-  }
-  await context.recordStep(stepIndex, startedAt, {
-    step: templateStep,
-    state: projected.execution.state,
-    resolvedActions: projected.execution.resolvedActions,
-    message: projected.execution.message,
-    artifacts: projected.execution.artifacts?.length
-      ? projected.execution.artifacts
-      : undefined,
-    diagnostics: projected.execution.diagnostics?.length
-      ? projected.execution.diagnostics.map((entry) =>
-          stampDiagnostic(entry, context.input, stepIndex, templateStep),
-        )
-      : undefined,
-    trace: projected.execution.trace?.length
-      ? projected.execution.trace
-      : undefined,
-  })
-  if (execution.replayDiverged) {
-    progress.replayDiverged = true
-    progress.state = 'failed'
-    progress.message = projected.execution.message
-    return false
-  }
-  if (execution.state === 'passed') return true
-  progress.state = execution.state
-  progress.message = projected.execution.message
-  return false
-}
-
-async function executeTargetSessionSteps(
-  session: TargetSession,
-  scenarioWallStartedAt: number,
-  context: SessionExecutionContext,
-): Promise<void> {
-  if (Boolean(session.executeScenario) === Boolean(session.executeStep)) {
-    recordExecutionError(
-      context.progress,
-      new Error(
-        'Target session must provide exactly one of executeStep or executeScenario',
-      ),
-      context.bindings,
-      context.input,
-      context.latestOccurredAt(),
-      undefined,
-    )
-    return
-  }
-  if (session.executeScenario) {
-    await executeScenarioSession(session, context)
-    return
-  }
-  await executeStepSession(session, scenarioWallStartedAt, context)
-}
-
-async function completeTargetSession(
-  session: TargetSession,
-  context: SessionExecutionContext,
-): Promise<TargetSessionCompletion | undefined> {
-  if (
-    context.progress.state === 'cancelled' ||
-    context.input.signal?.aborted ||
-    !session.complete
-  ) {
-    return undefined
-  }
-  try {
-    return validateCompletion(await session.complete())
-  } catch (error) {
-    recordExecutionError(
-      context.progress,
-      error,
-      context.bindings,
-      context.input,
-      context.latestOccurredAt(),
-      undefined,
-    )
-    return undefined
-  }
-}
-
-async function closeTargetSession(
-  session: TargetSession,
-  context: SessionExecutionContext,
-): Promise<void> {
-  try {
-    await session.close()
-  } catch (error) {
-    recordExecutionError(
-      context.progress,
-      error,
-      context.bindings,
-      context.input,
-      context.latestOccurredAt(),
-      undefined,
-    )
-  }
-}
-
-async function executeTargetSession(
-  session: TargetSession,
-  scenarioWallStartedAt: number,
-  context: SessionExecutionContext,
-): Promise<TargetSessionCompletion | undefined> {
-  try {
-    await executeTargetSessionSteps(session, scenarioWallStartedAt, context)
-    if (
-      context.progress.state === 'infrastructure-error' ||
-      context.progress.state === 'cancelled'
-    ) {
-      return undefined
-    }
-    return await completeTargetSession(session, context)
-  } finally {
-    await closeTargetSession(session, context)
-  }
-}
-
-function recordExecutionError(
-  progress: AttemptProgress,
-  error: unknown,
-  bindings: readonly ScenarioVariableBinding[],
-  input: ScenarioAttemptInput,
-  occurredAt: string,
-  signal?: AbortSignal,
-  stepIndex?: number,
-): void {
-  const attemptProgress = progress
-  const rawMessage = errorMessage(error)
-  attemptProgress.runtimeValueExposed ||= stringContainsBinding(
-    rawMessage,
-    bindings,
-  )
-  attemptProgress.state = isCancellation(error, signal)
-    ? 'cancelled'
-    : 'infrastructure-error'
-  attemptProgress.message = redactString(rawMessage, bindings)
-  attemptProgress.diagnostics.push(
-    runnerDiagnostic(
-      input,
-      attemptProgress.message,
-      occurredAt,
-      stepIndex,
-      stepIndex === undefined
-        ? undefined
-        : templateStepAt(input.scenario, stepIndex),
-    ),
-  )
-}
-
-async function executeScenarioSession(
-  session: ScenarioTargetSession,
-  context: SessionExecutionContext,
-): Promise<void> {
-  const { bindings, emit, input, progress, recordExecution } = context
-  try {
-    const scenarioExecution = await executeWithDeadline(
-      (operationSignal) => session.executeScenario(operationSignal),
-      input.signal,
-      input.timeout?.scenarioMs,
-      `Scenario exceeded its ${input.timeout?.scenarioMs}ms deadline`,
-    )
-    if (
-      scenarioExecution.stepExecutions.length > input.scenario.steps.length ||
-      (!scenarioExecution.replayDiverged &&
-        scenarioExecution.stepExecutions.length !== input.scenario.steps.length)
-    ) {
-      throw new Error(
-        'Scenario execution must return one result for every Scenario step',
-      )
-    }
-    for (const [
-      stepIndex,
-      execution,
-    ] of scenarioExecution.stepExecutions.entries()) {
-      const started = await emit({
-        type: 'step-started',
-        step: templateStepAt(input.scenario, stepIndex),
-        ...attemptIdentity(input, stepIndex),
-      })
-      if (!(await recordExecution(stepIndex, started.occurredAt, execution)))
-        break
-    }
-    if (scenarioExecution.replayDiverged) {
-      progress.replayDiverged = true
-      progress.state = 'failed'
-      progress.message ??= 'Replay diverged from the deterministic Scenario'
-    } else if (input.mode === 'replay') {
-      progress.replayedStepCount = input.scenario.steps.length
-    }
-  } catch (error) {
-    recordExecutionError(
-      progress,
-      error,
-      bindings,
-      input,
-      context.latestOccurredAt(),
-      input.signal,
-    )
-  }
-}
-
-async function executeEvaluatedStep(input: {
-  session: StepTargetSession
-  scenarioStartedAt: number
-  context: SessionExecutionContext
-  stepIndex: number
-  step: ScenarioStep
-  templateStep: ScenarioStep
-  evaluation: StepEvaluation
-  startedAt: string
-  recordAction: (input: ActionEvidenceInput) => Promise<ActionEvidence>
-}): Promise<StepExecution | undefined> {
-  const { bindings, progress, recordStep } = input.context
-  try {
-    const deadline = stepDeadline(
-      input.context.input.timeout,
-      input.scenarioStartedAt,
-    )
-    return await executeWithDeadline(
-      (operationSignal) =>
-        input.session.executeStep(input.step, operationSignal, {
-          stepIndex: input.stepIndex,
-          templateStep: input.templateStep,
-          runtimeBindings: input.context.input.scenario.runtimeBindings ?? [],
-          evaluation: input.evaluation,
-          recordAction: input.recordAction,
-        }),
-      input.context.input.signal,
-      deadline.timeoutMs,
-      deadline.timeoutMessage,
-    )
-  } catch (error) {
-    recordExecutionError(
-      progress,
-      error,
-      bindings,
-      input.context.input,
-      input.startedAt,
-      input.context.input.signal,
-      input.stepIndex,
-    )
-    await recordStep(input.stepIndex, input.startedAt, {
-      step: input.templateStep,
-      state: progress.state,
-      resolvedActions: [],
-      message: progress.message,
-    })
-    return undefined
-  }
-}
-
-function gherkinSourceEvidence(
-  input: ScenarioAttemptInput,
-  stepIndex: number,
-): ActionEvidence['source'] {
-  const step = templateStepAt(input.scenario, stepIndex)
-  return {
-    uri: input.specification.source.uri,
-    language: input.specification.source.language,
-    line: step.source?.line,
-    column: step.source?.column,
-    excerpt: step.source?.excerpt ?? `${step.keyword} ${step.text}`.trim(),
-  }
-}
-
-function createActionRecorder(
-  context: SessionExecutionContext,
-  stepIndex: number,
-): (input: ActionEvidenceInput) => Promise<ActionEvidence> {
-  let ordinal = 0
-  return async (input) => {
-    const actionOrdinal = ordinal++
-    const action = publicActionEvidence(
-      input,
-      {
-        id: `step-${stepIndex + 1}-action-${actionOrdinal + 1}`,
-        ordinal: actionOrdinal,
-        source: gherkinSourceEvidence(context.input, stepIndex),
-      },
-      context.bindings,
-    )
-    const emitted = await context.emit(
-      {
-        type: 'action-finished',
-        action,
-        ...attemptIdentity(context.input, stepIndex),
-      },
-      action.finishedAt,
-    )
-    return emitted.type === 'action-finished' ? emitted.action : action
-  }
-}
-
-async function advanceGapStep(input: {
-  session: StepTargetSession
-  scenarioStartedAt: number
-  context: SessionExecutionContext
-  stepIndex: number
-  step: ScenarioStep
-  cursor: GapCursor
-  cachePolicy: ExecutionCachePolicy
-}): Promise<{ cursor: GapCursor; stop: boolean }> {
-  const { context, stepIndex } = input
-  const { emit, input: attempt, progress, recordExecution } = context
-  if (attempt.signal?.aborted) {
-    progress.state = 'cancelled'
-    progress.message = 'Scenario cancelled before the next step started'
-    return { cursor: input.cursor, stop: true }
-  }
-  const templateStep = templateStepAt(attempt.scenario, stepIndex)
-  const started = await emit({
-    type: 'step-started',
-    step: templateStep,
-    ...attemptIdentity(attempt, stepIndex),
-  })
-  let evaluation = evaluationAt(input.cursor, stepIndex)
-  if (evaluation === 'adaptive') progress.adaptiveEvaluated = true
-  const stepInput = {
-    session: input.session,
-    scenarioStartedAt: input.scenarioStartedAt,
-    context,
-    stepIndex,
-    step: input.step,
-    templateStep,
-    startedAt: started.occurredAt,
-    recordAction: createActionRecorder(context, stepIndex),
-  }
-  let execution = await executeEvaluatedStep({ ...stepInput, evaluation })
-  if (!execution) return { cursor: input.cursor, stop: true }
-  let cursor = input.cursor
-  if (
-    evaluation === 'replay' &&
-    execution.replayDiverged &&
-    input.cachePolicy !== 'cache-only'
-  ) {
-    cursor = reseatGap(cursor, stepIndex)
-    evaluation = 'adaptive'
-    progress.adaptiveEvaluated = true
-    execution = await executeEvaluatedStep({ ...stepInput, evaluation })
-    if (!execution) return { cursor, stop: true }
-  }
-  const recorded = await recordExecution(
-    stepIndex,
-    started.occurredAt,
-    execution,
-  )
-  return { cursor, stop: !recorded }
-}
-
-async function executeStepSession(
-  session: StepTargetSession,
-  scenarioStartedAt: number,
-  context: SessionExecutionContext,
-): Promise<void> {
-  const { input, progress } = context
-  const adapter = input.adapter.executionCache
-  const prefix =
-    input.cacheEntry && adapter
-      ? cachedStepPrefixFrom(
-          input.cacheEntry,
-          input.scenario.steps.length,
-          adapter,
-        )
-      : undefined
-  let cursor = gapCursor(prefix)
-  const cachePolicy = input.cachePolicy ?? 'prefer-cache'
-  for (const [stepIndex, step] of input.scenario.steps.entries()) {
-    const advanced = await advanceGapStep({
-      session,
-      scenarioStartedAt,
-      context,
-      stepIndex,
-      step,
-      cursor,
-      cachePolicy,
-    })
-    cursor = advanced.cursor
-    if (advanced.stop) break
-  }
-  progress.replayedStepCount = cursor.replayUntil
 }
 
 function createAttemptEmitter(
