+++ packages/runner/src/results/test-run-id.ts
@@ -1,0 +1,11 @@
+export function validateTestRunId(id: string): void {
+  if (
+    !id ||
+    id === '.' ||
+    id === '..' ||
+    id.includes('/') ||
+    id.includes('\\')
+  ) {
+    throw new Error(`Invalid test run identifier "${id}"`)
+  }
+}
